+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 * fix: Unsilent in Eask-file by default (40d82becaf20f0851e5fc37aa17c5f6871c163a3)
 * Make better use for `special` commands (#206)
+* feat: Accept rest arguments after command line separator (#207)
 
 ## 0.9.x
 > Released Nov 17, 2023