+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 * feat(test): Add built-in `ecukes` test command (#236)
 * fix: Don't pollute outer programs (#239)
 * feat(generate/test): Add commands to setup test files (#243)
+* feat: Add `docs` command (#245)
 
 ## 0.9.x
 > Released Nov 17, 2023